fix: 모집인원 로직 수정
